+++ api_service/seattle_flu_incidence_mapper/models/generic_model.py
@@ -1,4 +1,7 @@
+import json
 from datetime import datetime
+
+from marshmallow import pre_load, post_dump
 from sqlalchemy import String, Column, DateTime
 from seattle_flu_incidence_mapper.orm_config import get_session, get_declarative_base, ma
 
@@ -12,11 +15,24 @@ class GenericModel(base):
     query_str = Column(String)
     model_type = Column(String)
     model_key = Column(String, primary_key=True)
-    rds_key = Column(String)
     created = Column(DateTime, default=datetime.utcnow)
 
 
 class GenericModelSchema(ma.ModelSchema):
+
+    @pre_load
+    def correct_query_str(self, *args, **kwargs):
+        pass
+
+    @post_dump
+    def correct_query_str(self, data, many=False):
+        if many and isinstance(data, list):
+            for i, d in enumerate(data):
+                data[i]['query_str'] = json.loads(d['query_str'])
+        else:
+            data['query_str'] = json.loads(data['query_str'])
+        return data
+
     class Meta:
         model = GenericModel
         sqla_session = get_session()