+++ api_service/HISTORY.rst
@@ -1,6 +1,12 @@
 History
 =======
 
+
+0.1.0
+=====
+Dropped embed model
+Added cached json models
+
 0.0.1
 =====
 Initial Project Skeleton added